backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
supabase>=2.3.0
python-dotenv>=1.0.0
pydantic>=2.6.0
httpx>=0.27.0
stripe>=8.0.0
fal-client>=0.2.0
boto3>=1.34.0
openai>=1.0.0
replicate>=0.25.0
python-slugify>=8.0.0
